handle case where residuals are empty
When not enough data are collected, or the fit is really bad, residuals
is empty (this doesn't appear to be documented, as far as I could tell).

This commit gracefully handles this case so the program doesn't crash
from an uncaught exception. I suppose I could have done a try-catch
instead of if-else...